+++ src/config/configmanager.cpp
@@ -358,6 +358,8 @@ bool ConfigManager::load() {
 	loadIntConfig(L, STAMINA_REGEN_MINUTE, "timeToRegenMinuteStamina", 3 * 60);
 	loadIntConfig(L, STAMINA_REGEN_PREMIUM, "timeToRegenMinutePremiumStamina", 6 * 60);
 	loadIntConfig(L, FAIRFIGHT_TIMERANGE, "fairFightTimeRange", 5 * 60 * 1000);
+	loadIntConfig(L, BLACK_SKULL_DEATH_HEALTH, "blackSkulledDeathHealth", 40);
+	loadIntConfig(L, BLACK_SKULL_DEATH_MANA, "blackSkulledDeathMana", 0);
 
 	loadStringConfig(L, CORE_DIRECTORY, "coreDirectory", "data");
 	loadStringConfig(L, DATA_DIRECTORY, "dataPackDirectory", "data-global");